requirements.txt
distlib==0.3.9
filelock==3.16.1
platformdirs==4.3.6
virtualenv==20.26.6
Pillow
requests
customtkinter
